requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pweave==0.23
PyYAML==3.11
Werkzeug==0.9.6
certifi==14.05.14
ipython==2.3.1
itsdangerous==0.24
pypandoc==0.8.4
pyzmq==14.4.1
tornado==4.0.2
